fix(ffe-icons): Remove svg-sprite dependency
BREAKING CHANGE: Since the tool used to build the sprite has changed
the `config` field in the configuration file is no longer available.

The sprite is no longer put in a `symbols/` folder automatically. If
you want to preserve the current folder structure you have to add
`symbols/` to `dest` in your config like so:

```diff
  {
-   "dest": "dist/assets/",
+   "dest": "dist/assets/symbol/",
    "icons": ["*"]
  }
```

This breaking change is introduced to get rid of PhantomJS from our
dependency tree. You should see a significant reduction in installation
time and disk usage from this change.