chore(ffe-icons-react): Upgrade to new ffe-icons
WIP

This version migrates to the new `ffe-icons`.

Some icons don't look right (they get clipped by the viewport or
something), so the transition to a new tool for `ffe-icons` didn't
work one hundred percent out of the box. Taking a break from this
now - I'll have to pick it up later.